view: view_builder: wait for schema agreement
This fixes a race in cql_test_env.cc. It starts the view_builder in

  view_builder->invoke_on_all(&db::view::view_builder::start).get();

And further down it has

  if (!env.local_db().has_keyspace(ks_name)) {
      env.create_keyspace(ks_name).get();
  }

The problem was that view_builder::start() would return immediately
with a ready future and start another future that would run
_mm.register_listener(this).

That registration could run at the same time that the keyspace
creation was in

  for (auto&& listener : _listeners) {

The addition of the new listener would reallocate _listeners and cause
a crash.

Note that this is effectively a revert of
17917e12ce86c5c97b2352a6a6507457919fa000.

Thanks to Avi for finding the bug.

Signed-off-by: Rafael Ávila de Espíndola <espindola@scylladb.com>
Message-Id: <20200115005402.97755-1-espindola@scylladb.com>